Update package declarations and imports after wordbreak package move (#1120)
