+++ querqy-core/src/main/java/querqy/rewrite/wordbreak/CompoundSuggestion.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package querqy.rewrite.contrib.wordbreak;
+package querqy.rewrite.wordbreak;
 
 import querqy.CharSequenceUtil;
