+++ querqy-core/src/main/java/querqy/rewrite/wordbreak/MorphologyProvider.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package querqy.rewrite.contrib.wordbreak;
+package querqy.rewrite.wordbreak;
 
 import java.util.HashMap;
 import java.util.Map;
