+++ querqy-core/src/main/java/querqy/rewrite/wordbreak/WordBreak.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package querqy.rewrite.contrib.wordbreak;
+package querqy.rewrite.wordbreak;
 
 import java.util.List;
 import java.util.Objects;
